[CHORE] 한 화면에서 SearchInput과 Output동작 구현
